Use a single class for the proxy and remove the base assets middleware
