+++ lib/shopify_cli/theme/dev_server/hot-reload-theme.js
@@ -88,7 +88,7 @@
     var sections = files.join(',');
     var expires = date.toUTCString();
 
-    document.cookie = `hot_reload_sections=${sections}; expires=${expires}; path=/`;
+    document.cookie = `hot_reload_files=${sections}; expires=${expires}; path=/`;
   }
 
   function refreshPage(files) {
