+++ test/shopify-cli/theme/dev_server/local_assets_test.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require "test_helper"
 require "shopify_cli/theme/dev_server"
 require "rack/mock"
@@ -99,7 +100,7 @@ module ShopifyCLI
           root = ShopifyCLI::ROOT + "/test/fixtures/theme"
           ctx = TestHelpers::FakeContext.new(root: root)
           theme = Theme.new(ctx, root: root)
-          stack = LocalAssets.new(ctx, app, theme: theme)
+          stack = LocalAssets.new(ctx, app, theme)
           request = Rack::MockRequest.new(stack)
           request.get(path)
         end
